# clientes-pedidos-service/src/main/java/com/multipedidos/clientes_pedidos_service/entity/Pedido.java
package com.multipedidos.clientes_pedidos_service.entity;

import jakarta.persistence.*;
import lombok.Data;

import java.util.List;

@Entity
@Table(name = "pedido")
@Data
public class Pedido {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @Column(nullable = false)
    private Long clienteId;

    private String nombre;

    private double subtotal;
    private double iva;
    private double descuentoPorcentaje;
    private double descuento;
    private double total;

    @ManyToMany(cascade = {CascadeType.PERSIST, CascadeType.MERGE})
    @JoinTable(
            name = "pedido_producto",
            joinColumns = @JoinColumn(name = "pedido_id"),
            inverseJoinColumns = @JoinColumn(name = "producto_id")
    )
    private List<Producto> productos;

    // 🔹 Campo de estado
    @Column(nullable = false)
    private String estado = "PENDIENTE"; // PENDIENTE, FACTURADO, ANULADO
}