# clientes-pedidos-service/src/main/java/com/multipedidos/clientes_pedidos_service/entity/Producto.java
package com.multipedidos.clientes_pedidos_service.entity;

import jakarta.persistence.*;
import lombok.Data;
import com.fasterxml.jackson.annotation.JsonBackReference;
import com.fasterxml.jackson.annotation.JsonIgnore;

import java.util.ArrayList;
import java.util.List;

@Entity
@Table(name = "producto")
@Data
public class Producto {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    private String nombre;
    private double precio;
    private int cantidad; 

    @ManyToMany(mappedBy = "productos")
    @JsonBackReference
    @JsonIgnore
    private List<Pedido> pedidos = new ArrayList<>();
}